Moved heap code to ledger_secure_sdk_sys
