+++ ledger_device_sdk/Cargo.toml
@@ -21,12 +21,7 @@ num-traits = { version = "0.2.14", default_features = false }
 rand_core = { version = "0.6.3", default_features = false }
 zeroize = { version = "1.6.0", default_features = false }
 numtoa = "0.2.4"
-<<<<<<< HEAD
 const-zero = "0.1.1"
-=======
-embedded-alloc = { version = "0.5.1", optional = true }
-critical-section = { version = "1.1.2", optional = true }
->>>>>>> e69708d (Heap and global allocator)
 
 [features]
 speculos = []
@@ -34,4 +29,4 @@ pre1_54 = []
 lib_bagl = []
 nbgl = []
 ccid = []
-heap = ["dep:embedded-alloc", "dep:critical-section"]
+heap = ["ledger_secure_sdk_sys/heap"]
