+++ ledger_device_sdk/src/lib.rs
@@ -29,7 +29,6 @@ pub mod ui;
 
 pub mod uxapp;
 
-use core::mem::MaybeUninit;
 use core::panic::PanicInfo;
 
 /// In case of runtime problems, return an internal error and exit the app
@@ -56,43 +55,6 @@ macro_rules! set_panic {
         }
     };
 }
-
-#[cfg(feature = "heap")]
-use critical_section::RawRestoreState;
-#[cfg(feature = "heap")]
-use embedded_alloc::Heap;
-
-#[cfg(feature = "heap")]
-#[global_allocator]
-static HEAP: Heap = Heap::empty();
-
-#[cfg(feature = "heap")]
-struct CriticalSection;
-#[cfg(feature = "heap")]
-critical_section::set_impl!(CriticalSection);
-
-/// Default empty implementation as we don't have concurrency.
-#[cfg(feature = "heap")]
-unsafe impl critical_section::Impl for CriticalSection {
-    unsafe fn acquire() -> RawRestoreState {}
-    unsafe fn release(restore_state: RawRestoreState) {}
-}
-
-/// Initializes the heap memory for the global allocator.
-///
-/// The heap is stored in the stack, and has a fixed size.
-/// This method is called just before [sample_main].
-#[no_mangle]
-#[cfg(feature = "heap")]
-extern "C" fn heap_init() {
-    const HEAP_SIZE: usize = 8192;
-    static mut HEAP_MEM: [MaybeUninit<u8>; HEAP_SIZE] = [MaybeUninit::uninit(); HEAP_SIZE];
-    unsafe { HEAP.init(HEAP_MEM.as_ptr() as usize, HEAP_SIZE) }
-}
-
-#[no_mangle]
-#[cfg(not(feature = "heap"))]
-extern "C" fn heap_init() {}
 
 extern "C" {
     fn c_main();
